move chunkers into dedicated package
